[src] rename Azure 2/3 references
