+++ src/main/java/hudson/plugins/spotinst/cloud/AzureSpotinstCloud.java
@@ -6,7 +6,7 @@ import hudson.plugins.spotinst.api.infra.ApiResponse;
 import hudson.plugins.spotinst.api.infra.JsonMapper;
 import hudson.plugins.spotinst.common.Constants;
 import hudson.plugins.spotinst.model.azure.AzureGroupInstance;
-import hudson.plugins.spotinst.model.azure.AzureVmSizeEnum;
+import hudson.plugins.spotinst.model.azure.AzureScaleSetSizeEnum;
 import hudson.plugins.spotinst.repos.IAzureGroupRepo;
 import hudson.plugins.spotinst.repos.RepoManager;
 import hudson.plugins.spotinst.slave.SlaveInstanceDetails;
@@ -248,8 +248,8 @@ public class AzureSpotinstCloud extends BaseSpotinstCloud {
     private Integer getNumOfExecutors(String vmSize) {
         LOGGER.info(String.format("Determining the # of executors for instance type: %s", vmSize));
 
-        Integer         retVal     = 1;
-        AzureVmSizeEnum vmSizeEnum = AzureVmSizeEnum.fromValue(vmSize);
+        Integer               retVal     = 1;
+        AzureScaleSetSizeEnum vmSizeEnum = AzureScaleSetSizeEnum.fromValue(vmSize);
 
         if (vmSizeEnum != null) {
             retVal = vmSizeEnum.getExecutors();
