+++ src/main/java/hudson/plugins/spotinst/model/azure/AzureGroupStatus.java
@@ -11,15 +11,15 @@ import java.util.List;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class AzureV3GroupStatus {
+public class AzureGroupStatus {
     //region members
-    private String               status;
-    private List<AzureV3GroupVm> vms;
-    private String               description;
+    private String             status;
+    private List<AzureGroupVm> vms;
+    private String             description;
     //endregion
 
     //region getters & setters
-    public AzureV3GroupStatus() {
+    public AzureGroupStatus() {
         this.vms = new ArrayList<>();
     }
 
@@ -31,11 +31,11 @@ public class AzureV3GroupStatus {
         this.status = status;
     }
 
-    public List<AzureV3GroupVm> getVms() {
+    public List<AzureGroupVm> getVms() {
         return vms;
     }
 
-    public void setVms(List<AzureV3GroupVm> vms) {
+    public void setVms(List<AzureGroupVm> vms) {
         this.vms = vms;
     }
