+++ src/main/java/hudson/plugins/spotinst/model/azure/AzureGroupStatusResponse.java
@@ -5,5 +5,5 @@ import hudson.plugins.spotinst.api.infra.BaseItemsResponse;
 /**
  * Created by Shibel Karmi Mansour on 01/12/2020.
  */
-public class AzureV3GroupStatusResponse extends BaseItemsResponse<AzureV3GroupStatus> {
+public class AzureGroupStatusResponse extends BaseItemsResponse<AzureGroupStatus> {
 }
