+++ src/main/java/hudson/plugins/spotinst/model/azure/AzureScaleUpResponse.java
@@ -5,5 +5,5 @@ import hudson.plugins.spotinst.api.infra.BaseItemsResponse;
 /**
  * Created by Shibel Karmi Mansour on 06/12/2020.
  */
-public class AzureV3ScaleUpResponse extends BaseItemsResponse<AzureV3ScaleResultNewVm> {
+public class AzureScaleUpResponse extends BaseItemsResponse<AzureScaleUpResultNewVm> {
 }
